Renames and moves.
