+++ src/Features/Core/Portable/EmbeddedLanguages/RegularExpressions/RegexDiagnosticAnalyzer.cs
@@ -4,8 +4,9 @@ using System.Collections.Generic;
 using System.Threading;
 using Microsoft.CodeAnalysis.CodeStyle;
 using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices
+namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.RegularExpressions
 {
     /// <summary>
     /// Analyzer that reports diagnostics in strings that we know are regex text.
