+++ src/Features/Core/Portable/Microsoft.CodeAnalysis.Features.csproj
@@ -116,5 +116,8 @@
     <PackageReference Include="System.Threading.Tasks.Extensions" Version="$(SystemThreadingTasksExtensionsVersion)" />
     <PackageReference Include="Microsoft.DiaSymReader" Version="$(MicrosoftDiaSymReaderVersion)" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="EmbeddedLanguages\LanguageServices\" />
+  </ItemGroup>
   <Import Project="..\..\..\Compilers\Core\AnalyzerDriver\AnalyzerDriver.projitems" Label="Shared" />
 </Project>
